refactor the scheduling endoints to make it easy to test
