+++ functions/actions/intreview/deleteEvent.js
@@ -1,0 +1,14 @@
+const { deleteEvent } = require("../../GoogleCalendar");
+module.exports = async (req, res) => {
+  // Delete an event with eventID
+  try {
+    if ("eventId" in req.body) {
+      const done = await deleteEvent(req.body.eventId);
+      return res.status(200).json({ done });
+    }
+  } catch (err) {
+    console.log({ err });
+    return res.status(400).json({ err });
+  }
+  return res.status(200).json({ done: false });
+};
